+++ poulpy-core/src/lib.rs
@@ -78,6 +78,7 @@ mod operations;
 mod scratch;
 mod utils;
 
+pub mod bench_suite;
 pub use operations::*;
 pub mod layouts;
 pub use automorphism::*;