+++ poulpy-hal/src/oep/svp_ppol.rs
@@ -47,25 +47,6 @@ pub unsafe trait SvpApplyDftToDftImpl<B: Backend> {
 
 /// # THIS TRAIT IS AN OPEN EXTENSION POINT (unsafe)
 /// * See the [poulpy-backend/src/cpu_fft64_ref/svp.rs](https://github.com/phantomzone-org/poulpy/blob/main/poulpy-backend/src/cpu_fft64_ref/svp.rs) reference implementation.
-/// * See [crate::api::SvpApplyDftToDftAdd] for corresponding public API.
-/// # Safety [crate::doc::backend_safety] for safety contract.
-pub unsafe trait SvpApplyDftToDftAddImpl<B: Backend> {
-    fn svp_apply_dft_to_dft_add_impl<R, A, C>(
-        module: &Module<B>,
-        res: &mut R,
-        res_col: usize,
-        a: &A,
-        a_col: usize,
-        b: &C,
-        b_col: usize,
-    ) where
-        R: VecZnxDftToMut<B>,
-        A: SvpPPolToRef<B>,
-        C: VecZnxDftToRef<B>;
-}
-
-/// # THIS TRAIT IS AN OPEN EXTENSION POINT (unsafe)
-/// * See the [poulpy-backend/src/cpu_fft64_ref/svp.rs](https://github.com/phantomzone-org/poulpy/blob/main/poulpy-backend/src/cpu_fft64_ref/svp.rs) reference implementation.
 /// * See [crate::api::SvpApplyDftToDftInplace] for corresponding public API.
 /// # Safety [crate::doc::backend_safety] for safety contract.
 pub unsafe trait SvpApplyDftToDftInplaceImpl: Backend {